Create TreatmentLog
This file basically just takes the TreatedCases and appends it to a log. This log is is just a linked list that holds all the cases. it has the size method, an append method for adding new treated cases. the asListOldestFirst method that returns a copy of the log in order of the oldest treatest case to newest treated case. and asListNewestFirst that returns a copy of the log in reverse order from Newest to Oldest treated case.